refactor: reorganize use case imports
