+++ app/routes/staff/analytics/csv-format/index.tsx
@@ -3,8 +3,8 @@ import { createRoute } from "honox/factory"
 import ChevronLeftIcon from "../../../../components/icons/lucide/chevronLeftIcon"
 import Callout from "../../../../components/ui/callout"
 import LinkButton from "../../../../components/ui/linkButton"
-import { ORDER_HISTORY_COLUMNS } from "../../../../usecases/exportOrderHistoryCsv"
-import { PRODUCT_CATALOG_COLUMNS } from "../../../../usecases/exportProductCatalogCsv"
+import { ORDER_HISTORY_COLUMNS } from "../../../../usecases/queries/exportOrderHistoryCsv"
+import { PRODUCT_CATALOG_COLUMNS } from "../../../../usecases/queries/exportProductCatalogCsv"
 import Layout from "../../-components/layout"
 
 type ColumnDescriptionProps = {
