+++ app/routes/staff/orders/[id]/delete/index.tsx
@@ -5,8 +5,8 @@ import Button from "../../../../../components/ui/button"
 import Callout from "../../../../../components/ui/callout"
 import LinkButton from "../../../../../components/ui/linkButton"
 import { setToastCookie } from "../../../../../helpers/ui/toast"
-import { getOrderDeletePageData } from "../../../../../usecases/getOrderDeletePageData"
-import { removeOrder } from "../../../../../usecases/removeOrder"
+import { removeOrder } from "../../../../../usecases/commands/removeOrder"
+import { getOrderDeletePageData } from "../../../../../usecases/queries/getOrderDeletePageData"
 import Layout from "../../../-components/layout"
 import OrderSummary from "../-components/orderSummary"
