+++ app/routes/staff/orders/[id]/edit/index.tsx
@@ -1,8 +1,8 @@
 import { validator } from "hono/validator"
 import { createRoute } from "honox/factory"
 import { setToastCookie } from "../../../../../helpers/ui/toast"
-import { getOrderEditPageData } from "../../../../../usecases/getOrderEditPageData"
-import { setOrderDetails } from "../../../../../usecases/setOrderDetails"
+import { setOrderDetails } from "../../../../../usecases/commands/setOrderDetails"
+import { getOrderEditPageData } from "../../../../../usecases/queries/getOrderEditPageData"
 import Layout from "../../../-components/layout"
 import OrderSummary from "../-components/orderSummary"
 import OrderEditForm from "./-components/$orderEditForm"
