+++ app/routes/staff/products/[id]/delete/index.tsx
@@ -5,8 +5,8 @@ import Button from "../../../../../components/ui/button"
 import Chip from "../../../../../components/ui/chip"
 import LinkButton from "../../../../../components/ui/linkButton"
 import { setToastCookie } from "../../../../../helpers/ui/toast"
-import { getProductDeletePageData } from "../../../../../usecases/getProductDeletePageData"
-import { removeProduct } from "../../../../../usecases/removeProduct"
+import { removeProduct } from "../../../../../usecases/commands/removeProduct"
+import { getProductDeletePageData } from "../../../../../usecases/queries/getProductDeletePageData"
 import { formatCurrencyJPY } from "../../../../../utils/money"
 import Layout from "../../../-components/layout"
