+++ app/routes/staff/products/[id]/edit/index.tsx
@@ -1,8 +1,8 @@
 import { validator } from "hono/validator"
 import { createRoute } from "honox/factory"
 import { setToastCookie } from "../../../../../helpers/ui/toast"
-import { getProductEditPageData } from "../../../../../usecases/getProductEditPageData"
-import { registerProduct } from "../../../../../usecases/registerProduct"
+import { registerProduct } from "../../../../../usecases/commands/registerProduct"
+import { getProductEditPageData } from "../../../../../usecases/queries/getProductEditPageData"
 import Layout from "../../../-components/layout"
 import ProductRegistrationForm from "../../-components/$productRegistrationForm"
 import { parseProductRegistrationFormData } from "../../-helpers/parseProductRegistrationFormData"
